+++ src/modules/padCommand.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import { NO_ACTIVE_EDITOR } from "../consts";
 import { getExtensionSettings } from "../helpers/tptSettings";
-import { getPureSelections, getSelectionContent, replaceSelectionsWithLines, sortSelectionsByPosition } from "../helpers/vsCodeHelpers";
+import { getPureSelections, getSelectionContentWithoutNewlines, replaceSelectionsWithLines, sortSelectionsByPosition } from "../helpers/vsCodeHelpers";
 
 export enum  PadDirection {
 	Start,
@@ -76,7 +76,7 @@ export async function padInternal(editor: vscode.TextEditor, direction: PadDirec
 	const fillString = padString || settings.defaultPadString || " ";
 	
 	for (let i = 0, len = selections.length; i < len; i++) {
-		const selectionContent = getSelectionContent(editor, selections[i]);
+		const selectionContent = getSelectionContentWithoutNewlines(editor, selections[i]);
 		if (direction === PadDirection.Start) {
 			replacesBySelection.push([selectionContent.padStart(padLength, fillString)]);
 		} else {